Add {file_extension} and {FILE_EXTENSION} as replaceable strings
I've wanted this as a feature, so I decided to implement it myself.

'{file_extension}' will be replaced with the lowercased file extension.
'{FILE_EXTENSION}' will be replaced with the uppercased file extension.

"Editing a .{file_extension} file" will yield "Editing a .ts file", or whatever file extension is being edited.